# bigint: allow `Sub` to work in-place on the RHS