@homu 2016-07-10
:pushpin: Commit 388a313 has been approved by `hauleth`

<!-- @homu r=hauleth 388a3132b8bcc7c5200bcdbd4889ad04f822ba0a -->