@homu 2016-07-10
:zap: Test exempted - [status](https://travis-ci.org/rust-num/num/builds/143481757)